chore: move composable theming to core ui module
