+++ .idea/misc.xml
@@ -8,6 +8,7 @@
         <entry key="../../../../layout/compose-model-1643704607731.xml" value="0.6173285198555957" />
         <entry key="../../../../layout/compose-model-1643719083688.xml" value="0.3935018050541516" />
         <entry key="../../../../layout/compose-model-1643733273751.xml" value="1.7761732851985559" />
+        <entry key="../../../../layout/compose-model-1643738306880.xml" value="1.7761732851985559" />
       </map>
     </option>
   </component>
