+++ app/src/main/java/com/messiaslima/promogamer/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.messiaslima.promogamer.ui.theme.PromoGamerTheme
+import com.messiaslima.promogamer.core.ui.theme.PromoGamerTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
