+++ buildSrc/src/main/kotlin/Kotlin.kt
@@ -1,4 +1,5 @@
 object Kotlin {
+    val gradlePlugin = GradlePlugin
     object GradlePlugin {
         const val androidVersion = "1.5.31"
         const val android = "org.jetbrains.kotlin.android"
