+++ buildSrc/src/main/kotlin/Module.kt
@@ -1,0 +1,7 @@
+object Module {
+    object Core {
+        private const val path = ":core"
+
+        const val ui = "$path:ui"
+    }
+}
