+++ core/ui/src/main/java/com/messiaslima/promogamer/core/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.messiaslima.promogamer.ui.theme
+package com.messiaslima.promogamer.core.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
