+++ core/ui/src/main/java/com/messiaslima/promogamer/core/ui/theme/Theme.kt
@@ -1,14 +1,14 @@
-package com.messiaslima.promogamer.ui.theme
+package com.messiaslima.promogamer.core.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
-import com.messiaslima.promogamer.ui.theme.PromoGamerColors.Purple200
-import com.messiaslima.promogamer.ui.theme.PromoGamerColors.Purple500
-import com.messiaslima.promogamer.ui.theme.PromoGamerColors.Purple700
-import com.messiaslima.promogamer.ui.theme.PromoGamerColors.Teal200
+import com.messiaslima.promogamer.core.ui.theme.PromoGamerColors.Purple200
+import com.messiaslima.promogamer.core.ui.theme.PromoGamerColors.Purple500
+import com.messiaslima.promogamer.core.ui.theme.PromoGamerColors.Purple700
+import com.messiaslima.promogamer.core.ui.theme.PromoGamerColors.Teal200
 
 private val DarkColorPalette = darkColors(
     primary = Purple200,
