+++ script/module/core.gradle
@@ -1,0 +1,39 @@
+apply plugin: Android.gradlePlugin.library
+apply plugin: Kotlin.gradlePlugin.android
+
+apply from: "$rootDir/script/detekt.gradle"
+
+android {
+    compileSdk = Android.compileSdk
+
+    defaultConfig {
+        minSdk = Android.minSdk
+        targetSdk = Android.targetSdk
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled = false
+            proguardFiles(
+                    getDefaultProguardFile("proguard-android-optimize.txt"),
+                    "proguard-rules.pro"
+            )
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_1_8
+        targetCompatibility = JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+
+dependencies {
+    testImplementation(Junit.junit)
+}
